#1052: epoch: combine all epoch manip into one file
