+++ src/vt/epoch/epoch_manip.cc
@@ -2,7 +2,7 @@
 //@HEADER
 // *****************************************************************************
 //
-//                              epoch_manip_make.cc
+//                                epoch_manip.cc
 //                           DARMA Toolkit v. 1.0.0
 //                       DARMA/vt => Virtual Transport
 //
@@ -185,6 +185,77 @@ void EpochManip::destroyScope(EpochScopeType scope) {
   // because if we did we would have to wait for termination of all epochs
   // within the scope before it could be destroyed. Thus, scopes can be quickly
   // created and destroyed by the user.
+}
+
+/*static*/ bool EpochManip::isRooted(EpochType const& epoch) {
+  constexpr BitPackerType::FieldType field = eEpochLayout::EpochIsRooted;
+  constexpr BitPackerType::FieldType size = 1;
+  return BitPackerType::boolGetField<field,size,EpochType>(epoch);
+}
+
+/*static*/ bool EpochManip::hasCategory(EpochType const& epoch) {
+  return BitPackerType::boolGetField<eEpochLayout::EpochHasCategory>(epoch);
+}
+
+/*static*/ eEpochCategory EpochManip::category(EpochType const& epoch) {
+  return BitPackerType::getField<
+    eEpochLayout::EpochCategory, epoch_category_num_bits, eEpochCategory
+  >(epoch);
+}
+
+/*static*/ NodeType EpochManip::node(EpochType const& epoch) {
+  return BitPackerType::getField<
+    eEpochLayout::EpochNode, node_num_bits, NodeType
+  >(epoch);
+}
+
+/*static*/ EpochType EpochManip::seq(EpochType const& epoch) {
+  return BitPackerType::getField<
+    eEpochLayout::EpochSequential, epoch_seq_num_bits, EpochType
+  >(epoch);
+}
+
+/*static*/ EpochScopeType EpochManip::getScope(EpochType const& epoch) {
+  // constexpr EpochScopeType offset = epoch_seq_num_bits - scope_limit;
+  return BitPackerType::getField<
+    eEpochLayout::EpochScope, scope_bits, EpochScopeType
+  >(epoch);
+}
+
+/*static*/
+void EpochManip::setIsRooted(EpochType& epoch, bool const is_rooted) {
+  BitPackerType::boolSetField<eEpochLayout::EpochIsRooted,1,EpochType>(epoch,is_rooted);
+}
+
+/*static*/
+void EpochManip::setHasCategory(EpochType& epoch, bool const has_cat) {
+  BitPackerType::boolSetField<eEpochLayout::EpochHasCategory,1,EpochType>(
+    epoch,has_cat
+  );
+}
+
+/*static*/
+void EpochManip::setScope(EpochType& epoch, EpochScopeType const scope) {
+  BitPackerType::setField<eEpochLayout::EpochScope,scope_bits>(epoch,scope);
+}
+
+/*static*/
+void EpochManip::setCategory(EpochType& epoch, eEpochCategory const cat) {
+  BitPackerType::setField<
+    eEpochLayout::EpochCategory, epoch_category_num_bits
+  >(epoch,cat);
+}
+
+/*static*/
+void EpochManip::setNode(EpochType& epoch, NodeType const node) {
+  BitPackerType::setField<eEpochLayout::EpochNode, node_num_bits>(epoch,node);
+}
+
+/*static*/
+void EpochManip::setSeq(EpochType& epoch, EpochType const seq) {
+  BitPackerType::setField<
+    eEpochLayout::EpochSequential, epoch_seq_num_bits
+  >(epoch,seq);
 }
 
 }} /* end namespace vt::epoch */
